Verschiedene Änderungen:
BeitragBox:

+++ src/de/hdm/gwt/itprojektws18/server/PinnwandVerwaltungImpl.java
@@ -550,25 +550,34 @@ public PinnwandVerwaltungImpl() {
 	@Override
 	public Like erstelleLike(Beitrag b, Nutzer n) {
 		
+		Like l = new Like();
 		/*
 		 * Überprüfen ob bereits ein Like-Objekt mit diesen Werten besteht
 		 */
-		Like l = getLikeFor(b.getId(), n.getId());
+		if (b != null && n != null) {
+			l = getLikeFor(b.getId(), n.getId());
+		} else {
+			return null;
+		}
+		
 		
 		/*
 		 * Gibt es noch kein Like-Objekt für diese Werte, so wird eines erstellt.
 		 */
-		if (l.getId() == 0) {
+		if (l == null) {
 			
 			Like like = new Like();
-			like.setBeitragFK(b.getId());
-			like.setErstellZeitpunkt(new Timestamp(System.currentTimeMillis()));
-			like.setNutzerFK(n.getId());
+			
 			/*
 			 * Setzen einer vorläufigen ID, welche nach Kommunikation mit der DB
 			 * auf den nächsthöchsten Wert gesetzt wird
 			 */
 			like.setId(1);
+			
+			like.setBeitragFK(b.getId());
+			like.setErstellZeitpunkt(new Timestamp(System.currentTimeMillis()));
+			like.setNutzerFK(n.getId());
+		
 			
 			return this.lMapper.insertLike(like);
 		} else {
@@ -583,7 +592,10 @@ public PinnwandVerwaltungImpl() {
 	 */
 	@Override
 	public void loeschen (Like l) {
-		this.lMapper.deleteLike(l);
+		if (l != null) {
+			this.lMapper.deleteLike(l);
+		}
+		
 	}
 	
 	/**
